Fix paragraph formatting
Vim now sets "formatexpr" when attaching an LSP to the current buffer.
This allows users to use the LSP's formatter when typing "gq".
However, this also breaks the old behavior of "gq" just hard-wrapping
the paragraph using text-width.

But we can use "gw" instead which behaves similarly but does not take
"formatexpr" into account.

See: https://vi.stackexchange.com/questions/39200/wrapping-comment-in-visual-mode-not-working-with-gq